imports: separate import from importer in tests
Break out common import test helpers into their own file, to avoid each
test file repeating the same code.
